fix: Model#canonical_link for submodels
The current implementation was always using the parent model's canonical link
... which is unstable (that is, the canonical link changes)

Gazebo itself does not have a notion of submodels, and even more so of canonical
links for submodels. This is purely a notion that is relevant for this library.
The new policy (first link of the submodel) makes more sense IMO